refactor: move serializers to a separate dir
