+++ openassessment/staffgrader/api.py
@@ -1,10 +1,10 @@
 """
 API endpoints for enhanced staff grader
 """
+from openassessment.staffgrader.serializers.submission_lock import SubmissionLockSerializer, TeamSubmissionLockSerializer
 from django.contrib.auth.decorators import login_required
 from django.http.response import HttpResponseBadRequest, HttpResponseForbidden, HttpResponseNotFound, HttpResponseServerError, JsonResponse
 from django.views.decorators.http import require_http_methods
-from rest_framework import serializers
 
 from openassessment.assessment.errors.staff import StaffAssessmentInternalError
 from openassessment.assessment.models.staff import StaffWorkflow, TeamStaffWorkflow
@@ -73,33 +73,3 @@ def locks_view(request):
         serializer = SubmissionLockSerializer(workflow)
 
     return JsonResponse(serializer.data)
-
-
-class SubmissionLockSerializer(serializers.ModelSerializer):
-    """
-    Create response payload with info about the workflow and operation success/failure
-    """
-    class Meta:
-        model = StaffWorkflow
-        fields = [
-            'submission_uuid',
-            'is_being_graded',
-            'grading_started_at',
-            'grading_completed_at',
-            'scorer_id'
-        ]
-
-
-class TeamSubmissionLockSerializer(serializers.ModelSerializer):
-    """
-    Create response payload with info about the workflow and operation success/failure
-    """
-    class Meta:
-        model = TeamStaffWorkflow
-        fields = [
-            'team_submission_uuid',
-            'is_being_graded',
-            'grading_started_at',
-            'grading_completed_at',
-            'scorer_id'
-        ]
